+++ src/components/behaviors/FarmingConfig.tsx
@@ -9,28 +9,29 @@ import ColorSelector from '../config/ColorSelector'
 
 import SlotBar from '../SlotBar'
 import { createSlotBars, FarmingConfigModel } from '../../models/BotConfig'
-import { useRef } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import { FrontendInfoModel } from '../../models/FrontendInfo'
 import Modal from '../Modal'
 import useModal from '../utils/UseModal'
 import YesNoModal from '../YesNoModal'
 import { MsFormat, StopWatchValues } from '../utils/StopWatch'
 import { DefaultValuesChecker } from '../utils/DefaultValuesChecker'
+import TimeInput from '../config/TimeInput'
 
 type Props = {
     className?: string,
     info: FrontendInfoModel | null,
     config: FarmingConfigModel,
     onChange: (config: FarmingConfigModel) => void,
-    running: boolean,
-    isCurrentMode: boolean,
     botStopWatch: StopWatchValues | null,
     botState: string,
 }
 
-const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMode, botStopWatch, botState }: Props) => {
+const FarmingConfig = ({ className, info, config, onChange, botStopWatch, botState }: Props) => {
     const statsModal = useModal()
     const debugModal = useModal()
+    const debugWarningModal = useModal(debugModal)
+    const onDeathModal = useModal(debugModal)
     const mobsNameDebugModal = useModal(debugModal)
     const mobsColorsDebugModal = useModal(mobsNameDebugModal)
     const resetSlotYesNo = useModal(debugModal)
@@ -50,6 +51,7 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
         'circle_pattern_rotation_duration': 30,
         'min_hp_attack': 30,
         'prevent_already_attacked': true,
+        'interval_between_buffs': 2000,
     }
 
     DefaultValuesChecker(config, defaultValues, onChange)
@@ -59,6 +61,12 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
         () => onChange({...config, ...{aggressive_mobs_colors: defaultValues['aggressive_mobs_colors'], aggressive_tolerence: defaultValues['aggressive_tolerence']} })
     ]
 
+    const [debugMode, setDebugMode] = useState(false);
+    useEffect(() => {
+        if (debugMode) {
+            debugWarningModal.open()
+        }
+    }, [debugMode])
     // StopWatchs
     const searchMobStopWatch = MsFormat(info?.last_search_duration ?? 0),
     fightStopWatch = MsFormat(info?.last_search_duration ?? 0)
@@ -67,21 +75,33 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
     const globalKPH = (Number(globalKPM) * 60).toFixed(2)
 
     return (
-        <>
+        <div className={className}>
             <SlotBar botMode="farming" config={config} onChange={onChange} />
             {/* DEBUG */}
             <YesNoModal isShowing={resetSlotYesNo.isShown} hide={resetSlotYesNo.close}
                 title={<h4>Confirm slot reset this action is irreversible</h4>}
                 onYes={() => {onChange({ ...config, slot_bars: createSlotBars() })}}/>
-            <Modal isShowing={debugModal.isShown} hide={debugModal.close} title={<h4>DEBUG</h4>} body={
+            <Modal isShowing={debugModal.isShown} hide={debugModal.close} title={<h4>Settings</h4>} body={
                 <ConfigTable>
                     <ConfigTableRow
-                        label={<ConfigLabel name="Mobs detection settings" helpText="" />}
+                        label={<ConfigLabel name="Mobs detection" helpText="" />}
                         item={<button onClick={mobsNameDebugModal.open}>⚙️</button>}
                     />
                     <ConfigTableRow
-                        label={<ConfigLabel name="Obstacle avoidance settings" helpText="" />}
+                        label={<ConfigLabel name="Avoidances" helpText="" />}
                         item={<button onClick={obstacleAvoidanceDebugModal.open}>⚙️</button>}
+                    />
+                    <ConfigTableRow
+                        label={<ConfigLabel name="On death event" helpText="" />}
+                        item={<button onClick={onDeathModal.open}>⚙️</button>}
+                    />
+                    <ConfigTableRow
+                        label={<ConfigLabel name="Interval between buffs" helpText="" />}
+                        item={<TimeInput value={config.interval_between_buffs} onChange={value => onChange({...config, interval_between_buffs: value})} />}
+                    />
+                    <ConfigTableRow
+                        label={<ConfigLabel name="Debug" />}
+                        item={<BooleanSlider value={debugMode} onChange={value => setDebugMode(value)} />}
                     />
                     <ConfigTableRow
                         label={<ConfigLabel name="Reset all slots" helpText="" />}
@@ -89,7 +109,7 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
                     />
                 </ConfigTable>
             }/>
-            <Modal isShowing={mobsColorsDebugModal.isShown} hide={mobsColorsDebugModal.close} title={(selectedMobType.current === 0)? <h4>Passive mob detection settings</h4> : <h4>Aggressive mob detection settings</h4>} body={
+            <Modal isShowing={mobsColorsDebugModal.isShown} hide={mobsColorsDebugModal.close} title={(selectedMobType.current === 0)? <h4>Passive mob detection</h4> : <h4>Aggressive mob detection</h4>} body={
                 <ConfigTable>
                     <ConfigTableRow
                         layout="v"
@@ -108,7 +128,7 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
                     />
                 </ConfigTable>
             }/>
-            <Modal isShowing={obstacleAvoidanceDebugModal.isShown} hide={obstacleAvoidanceDebugModal.close} title={<h4>Avoidances settings</h4>} body={
+            <Modal isShowing={obstacleAvoidanceDebugModal.isShown} hide={obstacleAvoidanceDebugModal.close} title={<h4>Avoidances</h4>} body={
                 <ConfigTable>
                     <ConfigTableRow
                         layout="v"
@@ -118,7 +138,7 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
                     <ConfigTableRow
                         layout="v"
                         label={<ConfigLabel name="Obstacle avoidance cooldown" helpText="Time before we try to move or escape if monster's HP doesn't change" />}
-                        item={<NumericInput unit='ms' value={config.obstacle_avoidance_cooldown} onChange={value => onChange({...config, obstacle_avoidance_cooldown: value})} />}
+                        item={<TimeInput value={config.obstacle_avoidance_cooldown} onChange={value => onChange({...config, obstacle_avoidance_cooldown: value})} />}
                     />
                     <ConfigTableRow
                         layout="v"
@@ -129,32 +149,35 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
             }/>
             <Modal isShowing={mobsNameDebugModal.isShown} hide={mobsNameDebugModal.close} title={<h4>Mobs detection</h4>} body={
                 <ConfigTable>
-                    <ConfigTableRow
-                        layout="v"
-                        label={<ConfigLabel name="Passive mob detection settings" helpText="" />}
-                        item={<button onClick={() => {
-                            selectedMobType.current = 0
-                            mobsColorsDebugModal.open()
-                        }}>⚙️</button>}
-                    />
-                    <ConfigTableRow
-                        layout="v"
-                        label={<ConfigLabel name="Agressive mob detection settings" helpText="" />}
-                        item={<button onClick={() => {
-                            selectedMobType.current = 1
-                            mobsColorsDebugModal.open()
-                        }}>⚙️</button>}
-                    />
-                    <ConfigTableRow
-                        layout="v"
-                        label={<ConfigLabel name="Min mobs name width" helpText="" />}
-                        item={<NumericInput unit='px' value={config.min_mobs_name_width} onChange={value => onChange({...config, min_mobs_name_width: value})} />}
-                    />
-                    <ConfigTableRow
-                        layout="v"
-                        label={<ConfigLabel name="Max mobs name width" helpText="" />}
-                        item={<NumericInput unit='px' value={config.max_mobs_name_width} onChange={value => onChange({...config, max_mobs_name_width: value})} />}
-                    />
+                    { debugMode &&
+                    <>
+                        <ConfigTableRow
+                            layout="v"
+                            label={<ConfigLabel name="Passive mob detection" helpText="" />}
+                            item={<button onClick={() => {
+                                selectedMobType.current = 0
+                                mobsColorsDebugModal.open()
+                            }}>⚙️</button>}
+                        />
+                        <ConfigTableRow
+                            layout="v"
+                            label={<ConfigLabel name="Agressive mob detection" helpText="" />}
+                            item={<button onClick={() => {
+                                selectedMobType.current = 1
+                                mobsColorsDebugModal.open()
+                            }}>⚙️</button>}
+                        />
+                        <ConfigTableRow
+                            layout="v"
+                            label={<ConfigLabel name="Min mobs name width" helpText="" />}
+                            item={<NumericInput unit='px' value={config.min_mobs_name_width} onChange={value => onChange({...config, min_mobs_name_width: value})} />}
+                        />
+                        <ConfigTableRow
+                            layout="v"
+                            label={<ConfigLabel name="Max mobs name width" helpText="" />}
+                            item={<NumericInput unit='px' value={config.max_mobs_name_width} onChange={value => onChange({...config, max_mobs_name_width: value})} />}
+                        />
+                    </> }
                     <ConfigTableRow
                         layout="v"
                         label={<ConfigLabel name="Circle pattern duration" helpText="The bot will try to move in a circle pattern to find targets. Value of 0 will stay in place. Lower the value to increase circle size. Default : 30" />}
@@ -169,12 +192,27 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
 
                 </ConfigTable>
             }/>
+            <Modal isShowing={debugWarningModal.isShown} hide={debugWarningModal.close}
+            title={<h4>WARNING - Debug mode </h4>} body={
+                <div className='modalCenteredContent'>
+                    <h1>Warning !</h1>
+                    <h2>For advanced users ONLY ! If something went wrong after you set something don't need to cry just reset your profile</h2>
+                </div>
+            }/>
             {/* DEBUG END */}
-{/*             <ConfigPanel>
+            <Modal isShowing={onDeathModal.isShown} hide={onDeathModal.close}
+            title={<h4>Auto disconnect</h4>} body={
                 <ConfigTable>
-
+                    <ConfigTableRow
+                        label={<ConfigLabel name="No mob time out" helpText="After this time character will disconnect if no mobs were found" />}
+                        item={<TimeInput value={config.mobs_timeout} onChange={value => onChange({...config, mobs_timeout: value})} />}
+                    />
+                    <ConfigTableRow
+                        label={<ConfigLabel name="Disconnect" helpText="Enable will automatically disconnect, otherwise we'll wait for revive" />}
+                        item={<BooleanSlider value={config.on_death_disconnect ?? true} onChange={value => onChange?.({ ...config, on_death_disconnect: value })} />}
+                    />
                 </ConfigTable>
-            </ConfigPanel> */}
+            }/>
             <Modal isShowing={statsModal.isShown} hide={statsModal.close}
             title={<h4>Stats - State: { botState }</h4>} body={
                 <div className="stats">
@@ -197,12 +235,6 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
                         <div>Global kills stats(approx): {globalKPM === "NaN" || globalKPM === "Infinity" ? 0 : globalKPM}/min
                         | {globalKPH === "NaN" || globalKPH === "Infinity" ? 0 : globalKPH}/hour</div>
                     </div>
-
-
-
-
-
-
                 </div>
             }/>
             {info && (
@@ -213,17 +245,22 @@ const FarmingConfig = ({ className, info, config, onChange, running, isCurrentMo
                     <div className="row">
                         <div>Target's detection mode: { config.is_stop_fighting? "🛑" : "✅" }</div>
                     </div>
-                    <button className="btn sm" onClick={statsModal.open}>Stats 📊</button>
                     <button className="btn sm"
-                        onClick={e => onChange?.({ ...config, is_stop_fighting: !config.is_stop_fighting })} >
-                        Detection 🎯
+                            onClick={e => onChange?.({ ...config, is_stop_fighting: !config.is_stop_fighting })} >
+                            Detection 🎯
                     </button>
-                    <button className="btn sm" onClick={debugModal.open}>Debug ⚙️</button>
+                    <div className="row">
+                        <button className="btn sm" onClick={statsModal.open}>Stats 📊</button>
+                        <button className="btn sm" onClick={debugModal.open}>Settings ⚙️</button>
+                    </div>
                 </div>
             )}
-        </>
+        </div>
     )
 }
 
 export default styled(FarmingConfig)`
+    & .modalCenteredContent {
+        text-align: center;
+    }
 `