+++ src/components/config/ConfigLabel.tsx
@@ -36,6 +36,7 @@ export default styled(ConfigLabel)`
     width: 100%;
     display: flex;
     flex-direction: column;
+    justify-content: space-between;
     gap: .5rem;
     color: white;