+++ src/models/BotConfig.tsx
@@ -91,14 +91,17 @@ export type FarmingConfigModel = Partial<{
     max_mobs_name_width: number,
 
     min_hp_attack: number,
-
-
+    on_death_disconnect: boolean,
+    interval_between_buffs: number,
+    mobs_timeout: number,
 }>
 
 export type SupportConfigModel = Partial<{
     [key: string]: any;
     slot_bars: SlotBars,
-    jump_cooldown: number,
+    obstacle_avoidance_cooldown: number,
+    on_death_disconnect: boolean,
+    interval_between_buffs: number,
 }>
 
 export type ShoutConfigModel = Partial<{